Commit merge Features/learn branch to master branch

+++ src/main/java/com/huuluc/englearn/mapper/MediaMapper.java
@@ -1,6 +1,6 @@
 package com.huuluc.englearn.mapper;
 
-import com.huuluc.englearn.entity.Media;
+import com.huuluc.englearn.model.Media;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
 
@@ -12,7 +12,7 @@ public interface MediaMapper {
     List<Media> findAll();
 
     @Select("SELECT media_id, media_name, media_type_id, url FROM media WHERE media_id = #{mediaId}")
-    Media findById(short mediaId);
+    Media findById(int mediaId);
 
     @Select("SELECT media_id, media_name, media_type_id, url FROM media WHERE media_name = #{mediaName}")
     Media findByName(String mediaName);
